--- src/main/java/com/kodilla/ecommercee/CartController.java
@@ -1,4 +1,0 @@
-package com.kodilla.ecommercee;
-
-public class CartController {
-}
